Break userRegisterActions into controller and service with appropriate tests
